use 'Float64' instead of 'Float' for type normalization of a 'real'
Summary:
The important change is in `xplat/msys/tools/cql/gen_sql.c`, which is the latest file listed below

This was the original plan, per https://fb.quip.com/nZnBAFb1tCwW#XYNABAolGJw
It's a very straightforward diff, this just changes the type hashes for float properties. The server code uses 'Float64' so we need to match.
I'm quite sure ricomariani (who's on PTO this week) will be okay with me doing this :)
Still, this could wait because we aren't experimenting with any tables that have `real` columns, but better sooner than later

Reviewed By: paulvanderspek

Differential Revision: D40503398

fbshipit-source-id: 26f0e268671a842dec88d2472a1882ed0be5bd8d