precomputing the key columns, value columns, and not null columns storing them in a table_info node that is present only for tables
Summary:
When using a backed table the offsets for the keys in the blob should be the order that the columns appear in the primary key constraint if there is one, rather than the declared order.  This gives you a chance to get ordering of your choice within the blob for locality.

To avoid this scan, and to avoid fattening the sem node for this and other "table only things" sem gets a new table_info node present for tables with this kind of thing.  It has:

* key column count and index
* value columns (not pk) count and index
* not null columns count and index

All of these were previously calculated by filtering on each usage.

The existing list of indexes on a table and its type hash were moved to this node as well.

An additional error check was added so that expressions are not allowed in the pk of backed or backing tables.  In general this *is* allowed.

Finally, we use the order of keys and values in the arrays to create the key and value blobs in an insert statement and to get the column offset in a select or update.

Differential Revision: D40663517

fbshipit-source-id: 1a12c63c90d481d554063741cca33742411b322b