update iteration check based on non-physical table status
Summary:
If we have a non-physical table (ex: backed table) we early exit out of our manage_recreate_table function, so we cannot wait for that table's turn to take action on the recreate group.

This problem only introduced because of D40203594 (https://github.com/facebookincubator/CG-SQL/commit/9a692af5a6e765445ea7b65308f2c9908c2318f4) and because in the next diff we want to change the recreate_annotation sort order to be create order instead of drop order.

Reviewed By: ricomariani

Differential Revision: D40444143

fbshipit-source-id: d68080eb2837939001c26436d1d76b5b48039e75