Remove CQL0449
Summary:
Changing the recreate group or tables was not allowed before because the moved table would run as part of the new group, but we would have a stale copy of the table that could being FKing to tables being deleted.

This is no longer a problem because we now have a hierarchy for recreate groups based on FK dependencies. We also properly traverse our DAG and drop children first so this is no longer a possible problem.

Reviewed By: ricomariani

Differential Revision: D40424763

fbshipit-source-id: c60e5131bfb62ce599a2036b5197e9823cf4ea13