fix recreate comparator logic
Summary:
The recreate_comparator function is used to 1. traverse the sorted elements to emit CRC if statement checks and the associated upgrades and 2. to traverse the elements in manage_recreate_tables and compile SQL string arguments to cql_rebuild_recreate group. For 1. we need the recreate_comparator to sort the group_ordinal in create order and for 2. we need the recreate_comparator to sort the ordinal in drop order.

This diff is a fix for 1. If we don't have create order for group_ordinal, then the child groups would be emitting CRC checks first, and try to rebuild/recreate. Then the parent checks would come next, where the children would be dropped - and then they would never get recreated.

Sample paste for the fixed upgrade is here: P539236152

Reviewed By: ricomariani

Differential Revision: D40440579

fbshipit-source-id: 01b5aaf6b61c1c7bf8169ed034807348442f754d