type_check fun rewrite leaks memory if value is a primitive node
Summary:
`type_check(...)` func is builtin func that allow validation that a value match a specific type at compile time.
e.g:
```
cast(
   type_check(get_general_id(args) as integer<id_general>) not null
   as integer<id_task>
)
```
This is a guarded cast.
After the type check, the node has served its purpose and is erased from the tree; so it's invisible to downstream code generation.

Now `type_check()` call in your code gets rewrite in the ast at semantic type validation. The outcome of the rewrite strip everything except the value we're doing the type checking on.
e.g:
```
let a := type_check(args as integer not null);

is rewrite after semantic validation to: ==>

let a := args;
```

Rewrite process in CQL consist on replacing the value of an ast node with a new ast node.
Now `type_check(...)` is a `ast_node` type with left and right child. Most node in CQL are this type. But we also few primitive node type that does not have left and right child like: `ast_num`, `ast_num`, `ast_blob` type.

Rewrite `type_check(cast(int_lit as int not null))` -> `cast(int_lit as int not null)` should work fine with the current because of node type have left and right child.

Now rewriting `type_check(int_lit as int not null)` -> `int_lit` is a problem because `int_lit` is a primitive node that does not have left and right now.

This diff fix the issue by removing the rewrite and keep `type_check` not expression as it is during the semantic analytic phase of the compilation and we strip it and replace it with the expr value at the codegen phase.
e.g:  type_check(cast(int_lit as int not null))
- after semantic analytic we keep the expression as it is
```
-> type_check(cast(int_lit as int not null))
```
- At codegen phase we strip it and replace it with just the expr:
```
 -> int_lit
```

This produce the exact same codegen output (no binary size increase) as before except the bug related to the rewrite is not happening anymore

Differential Revision: D41711257

fbshipit-source-id: cf51054458d429631ac7c2f8cb55d63a626757a4