Introduce cql:alias_of attribute
Summary:
Provide a way for CQL to call declare procs or functions under a different name (and with potentially different types) with a new `cql:alias_of` attribute. This is intended to be used for forward declarations of "native functions" that are predefined outside of CQL that is expected to be callable by generated CQL code.

Currently this is only implemented for C codegen for now.

internal
This attribute will give us a consistent way to alias native ops in both C, Mill-Lua, and Mill-Kotlin codegen. Function declarations that are supposed to be an alias of another function should have this attribute. This diff implements the C codegen path (emit `#define` to set the aliased name to be the name of the underlying function name). Mill-Lua codegen for this is handled in D42550722.

Reviewed By: RaoulFoaleng

Differential Revision: D43034445

fbshipit-source-id: 19d94a6c10e02ccfaf5ff7c0554a00640e490978