Fix CQL rollback logic
Summary: In our pattern of ROLLBACK TRANSACTION ... RELEASE SAVEPOINT, the rollback can fail because of automatic rollbacks (https://www.sqlite.org/lang_transaction.html). In this case we need to still release the savepoint - CQL currently skips over the RELEASE and skips to error cleanup. There is no harm in proceeding with a failed ROLLBACK.

Differential Revision: D43201656

fbshipit-source-id: 79cd6c61beef75863a721705ffc65eb4332204e0