Fix missing handling of CQL_DATA_TYPE_OBJECT in cql_multinull
Summary:
The API contract of `cql_multinull` is that it expects to consume exactly `2 * count` arguments by calling `va_arg`. For each of `count` columns, we first call `va_arg` to get the column type, then call `va_arg` again to get a pointer to where the value should be stored. But if we fail to call `va_arg` in the second case, all bets are off.

This patches the immediate issue. We can deal with how to prevent a recurrence of this issue for hypothetical new data types in a follow-on diff.

Differential Revision: D43665518

fbshipit-source-id: c0e2a743a72a3f9b7738477c57d2bda761751aae