Fix missing handling of CQL_DATA_TYPE_OBJECT in cql_multinull
Summary:
The API contract of `cql_multinull` is that it expects to consume exactly `2 * count` arguments by calling `va_arg`. For each of `count` columns, we first call `va_arg` to get the column type, then call `va_arg` again to get a pointer to where the value should be stored. But if we fail to call `va_arg` in the second case, all bets are off and we end up with [this crash](https://www.internalfb.com/logview/details/facebook_lite_android_crashes/257818f2665c75e26ca1fb597a004304/?selected-logview-tab=latest).

This patches the immediate issue. We can deal with how to prevent a recurrence of this issue for hypothetical new data types in a follow-on diff.

Reviewed By: RaoulFoaleng

Differential Revision: D43704544

fbshipit-source-id: b386dff867d5fb6902b199c3eefe09da81d3a9c2