\"declareProcs\"[ \t]*:          { return DECLARE_PROCS; } \"declareFuncs\"[ \t]*:          { return DECLARE_FUNCS; } \"createsObject\"[ \t]*:         { return CREATES_OBJECT; } \"returnType\"[ \t]*:            { return RETURN_TYPE; }
The json_replacements.l file did not have the correct case for declareProcs and declareFuncs (it used snake_case for both).

CREATES_OBJECT was missing as was RETURN_TYPE

SCHEMA was already correct.

To make future maintenance a little easier I added a helper function that formats a tag with the funky quotes in the right places.

Since all the lines were dirty anyway I also sorted them.