Diagrams update and required json_replacements.l changes for corrections and two missing (#171)
Summary:
I started with a railroad refresh only but it turned out there were some other minor issues.

We have this in the lexer:

```
\"declareProcs\"[ \t]*:          { return DECLARE_PROCS; }
\"declareFuncs\"[ \t]*:          { return DECLARE_FUNCS; }
\"createsObject\"[ \t]*:         { return CREATES_OBJECT; }
\"returnType\"[ \t]*:            { return RETURN_TYPE; }
```
but we were emitting `declare_procs` and `declare_funcs` in the railroad.  So fixed that.  CREATES_OBJECT was missing as was RETURN_TYPE so I added those.  I made a helper for printing the tag so the .l file is cleaner.  Since I had touched every line at that point I also sorted them.

I verified that there was no change in output except for those exact things (see diff).

the output for SCHEMA which is also new was already correct, so no change needed there.

Then I regenerated the docs.  Since the grammar had changed I built all the generated things not just the railroads.

Pull Request resolved: https://github.com/facebookincubator/CG-SQL/pull/171

Reviewed By: RaoulFoaleng

Differential Revision: D43816123

Pulled By: vaishmantha

fbshipit-source-id: 4d2b9f8b351c7a753cafaac98cf3a4e879aadef5